+++ drivers/video/s6d16a0x.c
@@ -343,10 +343,10 @@ static const unsigned char SEQ_DISPOFF[] = {
 static void s6d16a0x_spi_write(unsigned char address, unsigned char command)
 {
 	if (address != DATA_ONLY)
-		spi_gpio_write(s6d16a0x, 0x0, address);
+		spi_gpio_write(s6d16a0x, (unsigned int) 0x0, (unsigned int) address);
 
 	if (command != COMMAND_ONLY)
-		spi_gpio_write(s6d16a0x, 0x1, command);
+		spi_gpio_write(s6d16a0x, (unsigned int) 0x1, (unsigned int) command);
 }
 
 static void s6d16a0x_panel_send_sequence(const unsigned char *wbuf, unsigned int size_cmd)
